Assignment_2

+++ asssignmentapp/models.py
@@ -3,5 +3,5 @@ from django.db import models
 
 # Create your models here.
 class Counting(models.Model):
-    like = models.PositiveBigIntegerField(default=0)
-    dislike = models.PositiveBigIntegerField(default=0)
+    like = models.IntegerField(default=0)
+    dislike = models.IntegerField(default=0)
